pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crystal-prop-bench"
version = "0.1.0"
description = "Materials property prediction with calibrated uncertainty and chemical domain-shift evaluation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "pymatgen>=2024.2.0",
    "matminer>=0.9.0",
    "mp-api>=0.39.0",
    "lightgbm>=4.3.0",
    "scikit-learn>=1.4.0",
    "joblib>=1.3.0",
    "shap>=0.45.0",
    "pandas>=2.2.0",
    "pandera>=0.18.0",
    "mlflow>=2.11.0",
    "matplotlib>=3.8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pandas-stubs>=2.2.0",
]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "matminer.*",
    "mp_api.*",
    "shap.*",
    "lightgbm.*",
    "mlflow.*",
    "pymatgen.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "network: tests that require network access (deselect with '-m not network')",
]
